Add files via upload
This project uses two Arduino Boards, Arduino Mega the master and the Arduino Uno with CNC shield mounted over it. These two boards performs I2C(Inter IC communication). Arduino Mega makes all the control decisions as per command provided by the user and sends a request character to Arduino Uno. Uno is programmed in a way that on receiving one of four characters it rotates the motor for mentioned period of time and returns the control back to Mega. Mega is responsible to provide signals to LCD, LED, Keypad and other sub components.